fix: move most layout to root
Only i18n route checking is in locale layout
